+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/rules/frostWalkerNoFreezing/FrostWalkerEnchantmentMixin.java
@@ -1,0 +1,46 @@
+/*
+ * This file is part of the Carpet Ayaka Addition project, licensed under the
+ * GNU General Public License v3.0
+ *
+ * Copyright (C) 2025  Calboot and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.ayakacraft.carpetayakaaddition.mixin.rules.frostWalkerNoFreezing;
+
+import com.ayakacraft.carpetayakaaddition.CarpetAyakaSettings;
+import com.ayakacraft.carpetayakaaddition.utils.ModUtils;
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import me.fallenbreath.conditionalmixin.api.annotation.Condition;
+import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
+import net.minecraft.enchantment.FrostWalkerEnchantment;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Restriction(require = @Condition(value = ModUtils.MC_ID, versionPredicates = "<1.21.1"))
+@Mixin(FrostWalkerEnchantment.class)
+public class FrostWalkerEnchantmentMixin {
+
+    @WrapMethod(method = "freezeWater")
+    private static void wrapFreezeWater(LivingEntity entity, World world, BlockPos blockPos, int level, Operation<Void> original) {
+        if (!CarpetAyakaSettings.frostWalkerNoFreezing) {
+            original.call(entity, world, blockPos, level);
+        }
+    }
+
+}