+++ changelog.md
@@ -1,4 +1,5 @@
 ### General
 
 - ***Added*** carpet rule `bedrockNoBlastResistance`
+- ***Added*** carpet rule `frostWalkerNoFreezing`